LPD-82343 Separate notification-framework tests from object-action tests
Move notification template CRUD, search, and queue-entry API tests into
tests/notification-web/main/, while keeping object-action integration
tests (action-triggered notifications) in tests/object-web/notification/.
